The code is working! But only for one thread now. We still have to build the parser

+++ src/Model/SleepCode.java
@@ -3,10 +3,11 @@ package Model;
 public class SleepCode extends Code{
 
 	public Status run(Object obj) throws Exception{
-		Task a = new Task("1", 100.);
+		Task a = new Task("1", 100);
 
 		try {
-			Thread.sleep((Integer) obj);
+			Thread.sleep((Long) obj);
+			//System.out.println("["+((Long) obj)+"]");
 		} catch (InterruptedException e) {
 			return Status.FAIL;
 		}
